Update API Project
Adds Rocket API helpers and starts process of adding utilities for all
the different methods. This is for allowing web-server based playing of
the game. Obviously using HttpClient and blocking to delay threads is
not very good, so that will be addressed. Still need to add protos and
generated code as well as round out client code.